+++ apps/browser/src/tools/popup/settings/about-page/about-page-v2.component.ts
@@ -3,9 +3,7 @@ import { Component } from "@angular/core";
 import { RouterModule } from "@angular/router";
 import { firstValueFrom } from "rxjs";
 
-import { DeviceType } from "@bitwarden/common/enums";
 import { EnvironmentService } from "@bitwarden/common/platform/abstractions/environment.service";
-import { PlatformUtilsService } from "@bitwarden/common/platform/abstractions/platform-utils.service";
 import { CenterPositionStrategy, DialogService, ItemModule } from "@bitwarden/components";
 import { TroubleshootingDialogComponent } from "@bitwarden/logging-angular";
 import { I18nPipe } from "@bitwarden/ui-common";
@@ -15,20 +13,6 @@ import { PopOutComponent } from "../../../../platform/popup/components/pop-out.c
 import { PopupHeaderComponent } from "../../../../platform/popup/layout/popup-header.component";
 import { PopupPageComponent } from "../../../../platform/popup/layout/popup-page.component";
 import { AboutDialogComponent } from "../about-dialog/about-dialog.component";
-
-const RateUrls = {
-  [DeviceType.ChromeExtension]:
-    "https://chromewebstore.google.com/detail/bitwarden-free-password-m/nngceckbapebfimnlniiiahkandclblb/reviews",
-  [DeviceType.FirefoxExtension]:
-    "https://addons.mozilla.org/en-US/firefox/addon/bitwarden-password-manager/#reviews",
-  [DeviceType.OperaExtension]:
-    "https://addons.opera.com/en/extensions/details/bitwarden-free-password-manager/#feedback-container",
-  [DeviceType.EdgeExtension]:
-    "https://microsoftedge.microsoft.com/addons/detail/jbkfoedolllekgbhcbcoahefnbanhhlh",
-  [DeviceType.VivaldiExtension]:
-    "https://chromewebstore.google.com/detail/bitwarden-free-password-m/nngceckbapebfimnlniiiahkandclblb/reviews",
-  [DeviceType.SafariExtension]: "https://apps.apple.com/app/bitwarden/id1352778147",
-};
 
 // FIXME(https://bitwarden.atlassian.net/browse/CL-764): Migrate to OnPush
 // eslint-disable-next-line @angular-eslint/prefer-on-push-component-change-detection
@@ -48,7 +32,6 @@ export class AboutPageV2Component {
   constructor(
     private dialogService: DialogService,
     private environmentService: EnvironmentService,
-    private platformUtilsService: PlatformUtilsService,
   ) {}
 
   about() {
@@ -59,18 +42,6 @@ export class AboutPageV2Component {
 
   troubleshoot() {
     TroubleshootingDialogComponent.open(this.dialogService);
-  }
-
-  async launchHelp() {
-    const confirmed = await this.dialogService.openSimpleDialog({
-      title: { key: "continueToHelpCenter" },
-      content: { key: "continueToHelpCenterDesc" },
-      type: "info",
-      acceptButtonText: { key: "continue" },
-    });
-    if (confirmed) {
-      await BrowserApi.createNewTab("https://bitwarden.com/help/");
-    }
   }
 
   async openWebVault() {
@@ -84,19 +55,6 @@ export class AboutPageV2Component {
       const env = await firstValueFrom(this.environmentService.environment$);
       const url = env.getWebVaultUrl();
       await BrowserApi.createNewTab(url);
-    }
-  }
-
-  async rate() {
-    const confirmed = await this.dialogService.openSimpleDialog({
-      title: { key: "continueToBrowserExtensionStore" },
-      content: { key: "continueToBrowserExtensionStoreDesc" },
-      type: "info",
-      acceptButtonText: { key: "continue" },
-    });
-    if (confirmed) {
-      const deviceType = this.platformUtilsService.getDevice();
-      await BrowserApi.createNewTab((RateUrls as any)[deviceType]);
     }
   }
 }